Añadido componente de explicacion y ajustadas la habilidades

+++ captain_brinetooth/captain_brinetooth/components/Animation.cpp
@@ -24,7 +24,7 @@ Animation::Animation(std::string id, Texture* tex, int rows, int cols, int frame
 	speedModifier = &auxSpeedModifier;
 }
 
-Animation::Animation(std::string id, Texture* tex, int rows, int cols, int frames, Uint32 framerate, int repeat, int* modifier, Vector2D anchor) :
+Animation::Animation(std::string id, Texture* tex, int rows, int cols, int frames, Uint32 framerate, int repeat, float* modifier, Vector2D anchor) :
 	id_(id), tex_(tex), frames_(frames), framerate_(framerate), repeat_(repeat), startfr_(0), endfr_(frames - 1), anchorPoint_(anchor) {	//startfr_ y endfr_ se inicializan por defecto en esta constructora
 	framewidth_ = tex->width() / cols;
 	frameheight_ = tex->height() / rows;
@@ -32,7 +32,7 @@ Animation::Animation(std::string id, Texture* tex, int rows, int cols, int frame
 	speedModifier = modifier;
 }
 
-Animation::Animation(std::string id, Texture* tex, int rows, int cols, int frames, Uint32 framerate, int repeat, int startfr, int endfr, int* modifier, Vector2D anchor) :
+Animation::Animation(std::string id, Texture* tex, int rows, int cols, int frames, Uint32 framerate, int repeat, int startfr, int endfr, float* modifier, Vector2D anchor) :
 	id_(id), tex_(tex), frames_(frames), framerate_(framerate), repeat_(repeat), startfr_(startfr), endfr_(endfr), anchorPoint_(anchor) {	//startfr_ y endfr_ se inicializan por defecto en esta constructora
 	
 	framewidth_ = tex->width() / cols;
